Refactored code to deploy algorithms
